# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/orders/":{"post":{"tags":["Orders"],"summary":"Create a new order with multiple items","description":"Create a new order with multiple line items.\n\n## Request Body\n- `items`: List of items to include in the order, each with an item_id and quantity\n\n## Returns\n- `order_id`: ID of the created order\n- `total_price`: Calculated total price (including tax)\n- `status`: Order status\n\n## Raises\n- `400 Bad Request`: If the order has no items\n- `404 Not Found`: If any of the requested items don't exist","operationId":"create_order_api_orders__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"400":{"description":"Invalid request data"},"404":{"description":"One or more items not found"},"500":{"description":"Server error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}":{"get":{"tags":["Orders"],"summary":"Get order details by ID","description":"Retrieve detailed information about a specific order.\n\n## Path Parameter\n- `order_id`: The unique identifier of the order\n\n## Returns\n- Full order details including ID, status, line items, and calculated total price\n\n## Raises\n- `404 Not Found`: If the order with the specified ID doesn't exist","operationId":"get_order_api_orders__order_id__get","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"The ID of the order to retrieve","title":"Order Id"},"description":"The ID of the order to retrieve"}],"responses":{"200":{"description":"Order details retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDetailResponse"}}}},"404":{"description":"Order not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}/pay":{"post":{"tags":["Orders"],"summary":"Simulate payment and mark order as paid","description":"Simulate payment for an order.\n\n## Path Parameters\n- **order_id**: Unique identifier of the order to be paid.\n\n## Process\n- Validates that the order exists and is in 'pending_payment' status.\n- Updates the order status to 'in_preparation' (simulating successful payment).\n- Publishes the 'OrderPaid' event to trigger subsequent processing.\n\n## Returns\n- An OrderStatusResponse containing a success message, the order ID, and the updated status.\n\n## Raises\n- 404 Not Found: If the order does not exist.\n- 400 Bad Request: If the order is not in 'pending_payment' state.","operationId":"pay_order_api_orders__order_id__pay_post","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"ID of the order to pay","title":"Order Id"},"description":"ID of the order to pay"}],"responses":{"200":{"description":"Order status updated to in_preparation and OrderPaid event published","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusResponse"}}}},"400":{"description":"Invalid state for payment"},"404":{"description":"Order not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}/ready-to-ship":{"post":{"tags":["Orders"],"summary":"Mark order as ready to ship","description":"Mark an order as ready to ship.\n\n## Path Parameters\n- **order_id**: Unique identifier of the order to update.\n\n## Process\n- Validates that the order exists and is currently in 'in_preparation' state.\n- Updates the order status to 'ready_to_ship'.\n- Publishes the 'OrderReadyToShip' event to notify the carrier for pickup.\n\n## Returns\n- An OrderStatusResponse containing a success message, the order ID, and the updated status.\n\n## Raises\n- 404 Not Found: If the order does not exist.\n- 400 Bad Request: If the order is not in 'in_preparation' state.","operationId":"mark_order_ready_to_ship_api_orders__order_id__ready_to_ship_post","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"ID of the order to mark as ready to ship","title":"Order Id"},"description":"ID of the order to mark as ready to ship"}],"responses":{"200":{"description":"Order status updated to ready_to_ship and OrderReadyToShip event published","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusResponse"}}}},"400":{"description":"Invalid state for marking as ready to ship"},"404":{"description":"Order not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}/shipped":{"post":{"tags":["Orders"],"summary":"Mark order as shipped","description":"Mark an order as shipped.\n\n## Path Parameters\n- **order_id**: Unique identifier of the order to be shipped.\n\n## Process\n- Validates that the order exists and is in 'ready_to_ship' state.\n- Updates the order status to 'shipped'.\n- Publishes the 'Shipped' event to notify the customer that the package is on the way.\n\n## Returns\n- An OrderStatusResponse containing a success message, the order ID, and the updated status.\n\n## Raises\n- 404 Not Found: If the order does not exist.\n- 400 Bad Request: If the order is not in 'ready_to_ship' state.","operationId":"mark_order_shipped_api_orders__order_id__shipped_post","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"ID of the order to mark as shipped","title":"Order Id"},"description":"ID of the order to mark as shipped"}],"responses":{"200":{"description":"Order status updated to shipped and Shipped event published","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusResponse"}}}},"400":{"description":"Invalid state for shipping"},"404":{"description":"Order not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}/delivered":{"post":{"tags":["Orders"],"summary":"Mark order as delivered","description":"Mark an order as delivered.\n\n## Path Parameters\n- **order_id**: Unique identifier of the order to update.\n\n## Process\n- Validates that the order exists and is currently in 'shipped' state.\n- Updates the order status to 'delivered'.\n- Publishes the 'Delivered' event to notify the customer of successful delivery.\n\n## Returns\n- An OrderStatusResponse containing a success message, the order ID, and the updated status.\n\n## Raises\n- 404 Not Found: If the order does not exist.\n- 400 Bad Request: If the order is not in 'shipped' state.","operationId":"mark_order_delivered_api_orders__order_id__delivered_post","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"ID of the order to mark as delivered","title":"Order Id"},"description":"ID of the order to mark as delivered"}],"responses":{"200":{"description":"Order status updated to delivered and Delivered event published","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusResponse"}}}},"400":{"description":"Invalid state for delivery"},"404":{"description":"Order not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/":{"get":{"tags":["Products"],"summary":"List all products with pagination","description":"Retrieve a paginated list of products with sorting and filtering options.\n\n## Query Parameters\n- `page`: Page number (starts at 1)\n- `page_size`: Number of items per page (1-100)\n- `sort_by`: Field to sort by (id, name, price, stock)\n- `sort_order`: Sort direction (asc, desc)\n- `min_price`: Filter by minimum price\n- `max_price`: Filter by maximum price\n- `in_stock`: Filter by stock availability\n- `search`: Search term for name and description\n\n## Returns\nA paginated list of products with total count information","operationId":"list_products_api_products__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number, starting from 1","default":1,"title":"Page"},"description":"Page number, starting from 1"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Number of items per page (1-100)","default":20,"title":"Page Size"},"description":"Number of items per page (1-100)"},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","description":"Field to sort by (id, name, price, stock)","default":"id","title":"Sort By"},"description":"Field to sort by (id, name, price, stock)"},{"name":"sort_order","in":"query","required":false,"schema":{"type":"string","description":"Sort order (asc, desc)","default":"asc","title":"Sort Order"},"description":"Sort order (asc, desc)"},{"name":"min_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number","minimum":0},{"type":"null"}],"description":"Minimum price filter","title":"Min Price"},"description":"Minimum price filter"},{"name":"max_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number","exclusiveMinimum":0},{"type":"null"}],"description":"Maximum price filter","title":"Max Price"},"description":"Maximum price filter"},{"name":"in_stock","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"Filter by stock availability","title":"In Stock"},"description":"Filter by stock availability"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string","minLength":2},{"type":"null"}],"description":"Search in name and description","title":"Search"},"description":"Search in name and description"}],"responses":{"200":{"description":"List of products successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductList"}}}},"400":{"description":"Invalid query parameters"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Products"],"summary":"Create a new product","description":"Create a new product in the inventory.\n\n## Request Body\n- `name`: Product name (required)\n- `description`: Optional product description\n- `price`: Product price (required, > 0)\n- `stock`: Available inventory (required, >= 0)\n\n## Returns\nThe created product with its assigned ID\n\n## Raises\n- `400 Bad Request`: If the product data is invalid\n- `409 Conflict`: If a product with the same name already exists","operationId":"add_product_api_products__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}}},"responses":{"201":{"description":"Product successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"400":{"description":"Invalid product data"},"409":{"description":"Product with this name already exists"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/{product_id}":{"get":{"tags":["Products"],"summary":"Get detailed product information","description":"Retrieve detailed information about a specific product.\n\n## Path Parameter\n- **product_id**: The unique identifier of the product\n\n## Returns\nDetailed product information including stock level and pricing\n\n## Raises\n- **404 Not Found**: If the product with the specified ID doesn't exist","operationId":"get_product_api_products__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"The ID of the product to retrieve","title":"Product Id"},"description":"The ID of the product to retrieve"}],"responses":{"200":{"description":"Product details successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"404":{"description":"Product not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Products"],"summary":"Update an existing product","description":"Update an existing product's information.\n\n## Path Parameter\n- `product_id`: The unique identifier of the product to update\n\n## Request Body\n- `name`: Updated product name\n- `description`: Updated product description\n- `price`: Updated product price\n- `stock`: Updated inventory level\n\n## Returns\nThe updated product information\n\n## Raises\n- `404 Not Found`: If the product doesn't exist\n- `409 Conflict`: If updating would create a name conflict","operationId":"update_product_api_products__product_id__put","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"The ID of the product to update","title":"Product Id"},"description":"The ID of the product to update"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}}},"responses":{"200":{"description":"Product successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"400":{"description":"Invalid product data"},"404":{"description":"Product not found"},"409":{"description":"Name conflict with existing product"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Products"],"summary":"Delete a product","description":"Delete a product from the inventory.\n\n## Path Parameter\n- `product_id`: The unique identifier of the product to delete\n\n## Returns\nNo content on successful deletion\n\n## Raises\n- `404 Not Found`: If the product doesn't exist\n- `400 Bad Request`: If the product cannot be deleted (e.g., referenced by orders)","operationId":"delete_product_api_products__product_id__delete","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"The ID of the product to delete","title":"Product Id"},"description":"The ID of the product to delete"}],"responses":{"204":{"description":"Product successfully deleted"},"404":{"description":"Product not found"},"400":{"description":"Cannot delete product with active orders"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LineItemCreate":{"properties":{"item_id":{"type":"integer","exclusiveMinimum":0.0,"title":"Item Id","description":"ID of the item to order"},"quantity":{"type":"integer","exclusiveMinimum":0.0,"title":"Quantity","description":"Quantity of items to order"}},"type":"object","required":["item_id","quantity"],"title":"LineItemCreate","example":{"item_id":1,"quantity":2}},"LineItemDetail":{"properties":{"item_id":{"type":"integer","title":"Item Id"},"name":{"type":"string","title":"Name"},"quantity":{"type":"integer","title":"Quantity"},"unit_price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Unit Price"},"subtotal":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Subtotal"}},"type":"object","required":["item_id","name","quantity","unit_price","subtotal"],"title":"LineItemDetail"},"OrderCreate":{"properties":{"items":{"items":{"$ref":"#/components/schemas/LineItemCreate"},"type":"array","minItems":1,"title":"Items","description":"Items to include in the order"}},"type":"object","required":["items"],"title":"OrderCreate","example":{"items":[{"item_id":1,"quantity":2},{"item_id":3,"quantity":1}]}},"OrderDetailResponse":{"properties":{"order_id":{"type":"integer","title":"Order Id"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"items":{"items":{"$ref":"#/components/schemas/LineItemDetail"},"type":"array","title":"Items"},"total_price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Price"}},"type":"object","required":["order_id","status","created_at","items","total_price"],"title":"OrderDetailResponse","example":{"created_at":"2025-03-20T14:30:00","items":[{"item_id":1,"name":"Product A","quantity":2,"subtotal":49.98,"unit_price":24.99},{"item_id":3,"name":"Product C","quantity":1,"subtotal":9.99,"unit_price":9.99}],"order_id":123,"status":"created","total_price":59.97}},"OrderResponse":{"properties":{"message":{"type":"string","title":"Message"},"order_id":{"type":"integer","title":"Order Id"},"total_price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Price"},"status":{"type":"string","title":"Status"}},"type":"object","required":["message","order_id","total_price","status"],"title":"OrderResponse","example":{"message":"Order created successfully","order_id":123,"status":"created","total_price":59.99}},"OrderStatusResponse":{"properties":{"message":{"type":"string","title":"Message"},"order_id":{"type":"integer","title":"Order Id"},"status":{"type":"string","title":"Status"},"tracking_info":{"anyOf":[{"$ref":"#/components/schemas/ShippingInfo"},{"type":"null"}]}},"type":"object","required":["message","order_id","status"],"title":"OrderStatusResponse","example":{"message":"Order successfully marked as shipped","order_id":123,"status":"shipped","tracking_info":{"carrier":"FedEx","tracking_number":"FX123456789US"}}},"ProductCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Product name (1-100 characters)","example":"Ergonomic Keyboard"},"description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Description","description":"Optional product description (max 500 characters)","example":"Mechanical keyboard with ergonomic design"},"price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"}],"title":"Price","description":"Product price (greater than 0)","example":99.99},"stock":{"type":"integer","minimum":0.0,"title":"Stock","description":"Available inventory quantity (0 or greater)","example":25}},"type":"object","required":["name","price","stock"],"title":"ProductCreate","example":{"description":"High-precision gaming mouse with adjustable DPI","name":"Gaming Mouse","price":49.99,"stock":100}},"ProductList":{"properties":{"items":{"items":{"$ref":"#/components/schemas/ProductResponse"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"pages":{"type":"integer","title":"Pages"}},"type":"object","required":["items","total","page","page_size","pages"],"title":"ProductList","example":{"items":[{"created_at":"2025-03-20T18:30:45.123Z","description":"High-precision gaming mouse with adjustable DPI","id":1,"name":"Gaming Mouse","price":49.99,"stock":100}],"page":1,"page_size":20,"pages":3,"total":42}},"ProductResponse":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Product name (1-100 characters)","example":"Ergonomic Keyboard"},"description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Description","description":"Optional product description (max 500 characters)","example":"Mechanical keyboard with ergonomic design"},"price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$","title":"Price","description":"Product price (greater than 0)","example":99.99},"stock":{"type":"integer","minimum":0.0,"title":"Stock","description":"Available inventory quantity (0 or greater)","example":25},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["name","price","stock","id","created_at"],"title":"ProductResponse","example":{"created_at":"2025-03-20T18:30:45.123Z","description":"High-precision gaming mouse with adjustable DPI","id":1,"name":"Gaming Mouse","price":49.99,"stock":100}},"ShippingInfo":{"properties":{"carrier":{"type":"string","title":"Carrier","description":"Shipping carrier name"},"tracking_number":{"type":"string","title":"Tracking Number","description":"Package tracking number"}},"type":"object","required":["carrier","tracking_number"],"title":"ShippingInfo","example":{"carrier":"FedEx","tracking_number":"FX123456789US"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}